requirements.txt
flask==2.2.3
Werkzeug==2.2.3
Jinja2==3.1.2
itsdangerous==2.1.2
click==8.1.3
MarkupSafe==2.1.2
PyPDF2==3.0.1
markdown==3.8
python-dotenv==1.0.0
anthropic==0.25.9
openai==1.35.3
requests==2.31.0
httpx==0.27.0 